Flush stdout after printing a progress message
Without this, messages "building XYZ" can appear at the end of the CI
log, after all error messages (if there are any). For example, when
building three packages pkg1, pkg2 and pkg3, and pkg3 fails with an
error message, the output can look like this:

    error: something failed in pkg3
    building pkg1
    building pkg2
    building pkg3

This can be quite confusing, because it appears that the error message
from pkg3 was printed before the build has actually started. The
reason for incorrect message ordering is that stdout (used for
"building XYZ" messages) is fully buffered if not connected to a
terminal. Error messages from child processes do not go through the
same buffer and can therefore appear earlier, if the child flushes its
buffer or writes to stderr, which is line-buffered.